backend/requirements.txt
Flask==2.2.3
Werkzeug<3.0
google-generativeai==0.3.2
python-dotenv==1.0.0